+++ packages/@expo/cli/src/start/server/metro/MetroBundlerDevServer.ts
@@ -10,6 +10,7 @@ import { BundlerDevServer, BundlerStartOptions, DevServerInstance } from '../Bun
 import { CreateFileMiddleware } from '../middleware/CreateFileMiddleware';
 import { HistoryFallbackMiddleware } from '../middleware/HistoryFallbackMiddleware';
 import { InterstitialPageMiddleware } from '../middleware/InterstitialPageMiddleware';
+import { ReactDevToolsPageMiddleware } from '../middleware/ReactDevToolsPageMiddleware';
 import {
   DeepLinkHandler,
   RuntimeRedirectMiddleware,
@@ -84,6 +85,7 @@ export class MetroBundlerDevServer extends BundlerDevServer {
         scheme: options.location.scheme ?? null,
       }).getHandler()
     );
+    middleware.use(new ReactDevToolsPageMiddleware(this.projectRoot).getHandler());
 
     const deepLinkMiddleware = new RuntimeRedirectMiddleware(this.projectRoot, {
       onDeepLink: getDeepLinkHandler(this.projectRoot),